# chore: Log level detection improvements