**What this PR does / why we need it**:
1. Add a benchmark for log-level detection code.
2. Check for log levels in standard logging formats before doing a broader search. This change has sped up the code to detect log formats from the logline based on the benchmark:
```
benchmark                                   old ns/op     new ns/op     delta
Benchmark_extractLogLevelFromLogLine-10     3427          3013          -12.08%
Benchmark_extractLogLevelFromLogLine-10     3443          3008          -12.63%
Benchmark_extractLogLevelFromLogLine-10     3498          3007          -14.04%
Benchmark_extractLogLevelFromLogLine-10     3496          3010          -13.90%
Benchmark_extractLogLevelFromLogLine-10     3504          3015          -13.96%

benchmark                                   old allocs     new allocs     delta
Benchmark_extractLogLevelFromLogLine-10     1              1              +0.00%
Benchmark_extractLogLevelFromLogLine-10     1              1              +0.00%
Benchmark_extractLogLevelFromLogLine-10     1              1              +0.00%
Benchmark_extractLogLevelFromLogLine-10     1              1              +0.00%
Benchmark_extractLogLevelFromLogLine-10     1              1              +0.00%

benchmark                                   old bytes     new bytes     delta
Benchmark_extractLogLevelFromLogLine-10     16            16            +0.00%
Benchmark_extractLogLevelFromLogLine-10     16            16            +0.00%
Benchmark_extractLogLevelFromLogLine-10     16            16            +0.00%
Benchmark_extractLogLevelFromLogLine-10     16            16            +0.00%
Benchmark_extractLogLevelFromLogLine-10     16            16            +0.00%
```

**Checklist**
- [x] Tests updated